src: upstream_patches_ui: Add 'Register/Unregister Mailing Lists' menu
The user has one chance to manage the registered mailing lists that is
when the feature is first launched. After that, the only way to
add/remove lists is by manually editing the `lore.config` file, which
is both impractical and error prone.

This commit adds a menu entry inside the 'Dashboard'->'Settings' named
'Register/Unresgister Mailing Lists' that allows the user to revisit the
screen mentioned above and set the desired mailing lists available in
lore.kernel.org.

Note that this commit also adds the menu 'Settings', as the addition or
removal of mailing lists is a setting and, when we need to add new menus
related to settings, this top menu is already in place.

Closes: #811

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>