src: lib: dialog_ui: Fix `create_simple_checklist` checkmark bug
The function defined in `dialog_ui.sh` that creates a dialog simple
checklist has the capability of marking the already checked entries of
the list. However, that capability has a bug results in false positive
(i.e., mark unchecked entries). The bug is due to the function testing
if the `check_statuses` (array used to determine the checked entries)
positions are set, not if they are set to 1 (the correct approach). This
resulted in positions set with anything, including 0 (which should mean
unchecked), to be checked.

This commit fixes this bug by making the appropriate test. It also
updates the test suite to replicate the conditions of the bug, i.e., set
a position in the `check_statuses` with 0 and assert that this entry is
unchecked.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>